requirements.txt
pysocks
requests
eth-account
web3
python-dotenv
web3-proxy-providers